refactor(lib): move getSocialLinks to a helpers file
